expose `write_legacy`

+++ src/iterative_telemetry/__init__.py
@@ -37,6 +37,7 @@ class IterativeTelemetryLogger:
         url=URL,
         token=TOKEN,
         debug: bool = False,
+        write_legacy: bool = True,
     ):
         self.tool_name = tool_name
         self.tool_version = tool_version
@@ -44,6 +45,7 @@ class IterativeTelemetryLogger:
         self.url = url
         self.token = token
         self.debug = debug
+        self.write_legacy = write_legacy
         if self.debug:
             logger.setLevel(logging.DEBUG)
             logger.debug("IterativeTelemetryLogger is in debug mode")
@@ -75,7 +77,8 @@ class IterativeTelemetryLogger:
         return (
             os.environ.get(DO_NOT_TRACK_ENV, None) is None and self.enabled()
             if callable(self.enabled)
-            else self.enabled and _find_or_create_user_id() is not None
+            else self.enabled
+            and _find_or_create_user_id(self.write_legacy) is not None
         )
 
     def send(
@@ -153,7 +156,7 @@ class IterativeTelemetryLogger:
             # "tool_source": self.tool_source, # TODO
             # "scm_class": _scm_in_use(),
             **_system_info(),
-            "user_id": _find_or_create_user_id(),
+            "user_id": _find_or_create_user_id(self.write_legacy),
             "group_id": "",  # TODO
         }
 
@@ -193,7 +196,7 @@ def generate_id():
 
 
 @lru_cache(None)
-def _find_or_create_user_id():
+def _find_or_create_user_id(write_legacy=True):
     """
     The user's ID is stored on a file under the global config directory.
     The file should contain JSON with a `user_id` key:
@@ -221,7 +224,11 @@ def _find_or_create_user_id():
 
             # only for non-DVC packages,
             # write legacy file in case legacy DVC is installed later
-            if not old.exists() and uid.lower() != DO_NOT_TRACK_VALUE.lower():
+            if (
+                write_legacy
+                and not old.exists()
+                and uid.lower() != DO_NOT_TRACK_VALUE.lower()
+            ):
                 json.dump({"user_id": uid}, old.open("w", encoding="utf8"))
 
             if uid.lower() != DO_NOT_TRACK_VALUE.lower():
